Chapter10/requirements.txt
appnope==0.1.0
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
bcrypt==3.2.0
bleach==3.1.5
cffi==1.14.2
click==7.1.2
constantly==15.1.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Script==2.0.6
Flask-Twisted==0.1.2
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.6.0
incremental==17.5.0
iniconfig==1.0.1
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.5.0
mysql-connector-python==8.0.20
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
observable==1.0.3
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.2
python-dateutil==2.8.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.17
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
zope.interface==5.1.1
